..F....... [ZBXNEXT-221] fixed triggers filter reset to maintain filter_hosts value, if single host selected
git-svn-id: svn://svn.zabbix.com/branches/dev@88612 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-221/frontends/php/triggers.php
@@ -648,7 +648,7 @@ else {
 		$filter_name = '';
 		$filter_priority = [];
 		$filter_groupids = [];
-		$filter_hostids = [];
+		$filter_hostids = getRequest('filter_hostids', CProfile::getArray('web.triggers.filter_hostids', []));
 		$filter_state = -1;
 		$filter_status = -1;
 		$filter_value = -1;
@@ -867,7 +867,9 @@ else {
 		CProfile::delete('web.triggers.filter_name');
 		CProfile::delete('web.triggers.filter_priority');
 		CProfile::delete('web.triggers.filter_groupids');
-		CProfile::delete('web.triggers.filter_hostids');
+		if (count($filter_hostids) != 1) {
+			CProfile::delete('web.triggers.filter_hostids');
+		}
 		CProfile::delete('web.triggers.filter_state');
 		CProfile::delete('web.triggers.filter_status');
 		CProfile::delete('web.triggers.filter.evaltype');
